# doc/swagger/default_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "瑞吉外卖接口文档",
    "version": "1.0",
    "title": "瑞吉外卖"
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/setmeal": {
      "post": {
        "tags": [
          "setmeal-controller"
        ],
        "summary": "save",
        "operationId": "saveUsingPOST_4",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "setmealDto",
            "description": "setmealDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SetmealDto",
              "originalRef": "SetmealDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»",
              "originalRef": "R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "setmeal-controller"
        ],
        "summary": "update",
        "operationId": "updateUsingPUT_4",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "setmealDto",
            "description": "setmealDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SetmealDto",
              "originalRef": "SetmealDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»",
              "originalRef": "R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "setmeal-controller"
        ],
        "summary": "delete",
        "operationId": "deleteUsingDELETE_3",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "required": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»",
              "originalRef": "R«string»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/setmeal/list": {
      "get": {
        "tags": [
          "setmeal-controller"
        ],
        "summary": "list",
        "operationId": "listUsingGET_3",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "createTime",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "createUser",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "image",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "isDeleted",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "price",
            "in": "query",
            "required": false,
            "type": "number"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "updateTime",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "updateUser",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«List«Setmeal»»",
              "originalRef": "R«List«Setmeal»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/setmeal/page": {
      "get": {
        "tags": [
          "setmeal-controller"
        ],
        "summary": "page",
        "operationId": "pageUsingGET_4",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "name",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«Page»",
              "originalRef": "R«Page»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/setmeal/status/{status}": {
      "post": {
        "tags": [
          "setmeal-controller"
        ],
        "summary": "status",
        "operationId": "statusUsingPOST_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "required": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "status",
            "in": "path",
            "description": "status",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»",
              "originalRef": "R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/setmeal/{id}": {
      "get": {
        "tags": [
          "setmeal-controller"
        ],
        "summary": "get",
        "operationId": "getUsingGET_2",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«SetmealDto»",
              "originalRef": "R«SetmealDto»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "OrderItem": {
      "type": "object",
      "properties": {
        "asc": {
          "type": "boolean",
          "refType": null
        },
        "column": {
          "type": "string",
          "refType": null
        }
      },
      "title": "OrderItem"
    },
...
    "Page": {
      "type": "object",
      "properties": {
        "countId": {
          "type": "string",
          "refType": null
        },
        "current": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "hitCount": {
          "type": "boolean",
          "refType": null
        },
        "maxLimit": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "optimizeCountSql": {
          "type": "boolean",
          "refType": null
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem",
            "originalRef": "OrderItem"
          },
          "refType": "OrderItem"
        },
        "pages": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "records": {
          "type": "array",
          "items": {
            "type": "object"
          },
          "refType": "object"
        },
        "searchCount": {
          "type": "boolean",
          "refType": null
        },
        "size": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "refType": null
        }
      },
      "title": "Page"
    },
...
    "R«List«Setmeal»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Setmeal",
            "originalRef": "Setmeal"
          }
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«List«Setmeal»»"
    },
...
    "R«Page»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/Page",
          "originalRef": "Page"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«Page»"
    },
    "R«SetmealDto»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/SetmealDto",
          "originalRef": "SetmealDto"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«SetmealDto»"
    },
...
    "R«string»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "string"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«string»"
    },
    "Setmeal": {
      "type": "object",
      "properties": {
        "categoryId": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "code": {
          "type": "string",
          "refType": null
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "description": {
          "type": "string",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "image": {
          "type": "string",
          "refType": null
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "name": {
          "type": "string",
          "refType": null
        },
        "price": {
          "type": "number",
          "refType": null
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "refType": null
        }
      },
      "title": "Setmeal"
    },
    "SetmealDish": {
      "type": "object",
      "properties": {
        "copies": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "dishId": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "name": {
          "type": "string",
          "refType": null
        },
        "price": {
          "type": "number",
          "refType": null
        },
        "setmealId": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "sort": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "refType": null
        }
      },
      "title": "SetmealDish"
    },
    "SetmealDto": {
      "type": "object",
      "properties": {
        "categoryId": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "categoryName": {
          "type": "string",
          "refType": null
        },
        "code": {
          "type": "string",
          "refType": null
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "description": {
          "type": "string",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "image": {
          "type": "string",
          "refType": null
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "name": {
          "type": "string",
          "refType": null
        },
        "price": {
          "type": "number",
          "refType": null
        },
        "setmealDishes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SetmealDish",
            "originalRef": "SetmealDish"
          },
          "refType": "SetmealDish"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "refType": null
        }
      },
      "title": "SetmealDto"
    },
...
  }
}